refactor: streets move out of CityGenerator (#11)
gen/Streets holds everything between "this chunk is a street" and the
blocks that make one: the surface, the slope sections where the city
level steps, the connectors to minor streets, the border walls and their
supports, and the front parts of the buildings facing them. Nineteen
methods, and in the source they were already one contiguous 442-line
block plus three stragglers.

Three helpers widened to public - setBlocksFromPalette, fillSupportBelow
and clearToMax. That follows the shape the package already has rather
than adding to it: generatePart, getRandomPart and getMinHeightAt are
public for exactly this reason, and are the vocabulary Highways, Railways
and Scattered already speak.

Pure code motion, verified token-for-token against the originals modulo
the added feature parameter: the street block is 2079 tokens on both
sides, doBorder and isHigherThenNearbyStreetChunk 63, isCorner 71.

The verification earned its keep here. Four separate blind-replacement
corruptions were caught getting to this point, the worst being a rewrite
of `isEmpty(` that turned every `list.isEmpty()` in the block into
`list.CityGenerator.isEmpty()`. That one happened to fail to compile;
the shape of the same mistake that does compile is a silently moved RNG
draw, which is why the token comparison exists rather than a reading.

HardAirSetting stays in CityGenerator - it sat inside the moved range but
Highways, Railways and Scattered all name it.

CityGenerator 1891 -> 1420 lines; 2425 at the start of this milestone.

All seven digest suites unchanged.
